moved Fidelity and Implementation enumerators from pylele_config to pylele_api
